Refac

+++ Server/src/controllers/deleteFav.js
@@ -3,10 +3,14 @@ const { Favorite } = require("../DB_connection");
 const deleteFav = async (req, res) => {
   const { id } = req.params;
   try {
-    await Favorite.destroy({ where: { id: id } });
-    const allFavs = await Favorite.findAll();
+    if (id) {
+      await Favorite.destroy({ where: { id: id } });
 
-    return res.json(allFavs);
+      const allFavs = await Favorite.findAll();
+      return res.json(allFavs);
+    }
+
+    return res.status(500).json({ error: "Faltan datos" });
   } catch (error) {
     return res.status(500).json(error.message);
   }
